requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cloudinary==1.41.0
Pillow==10.4.0
tensorflow==2.16.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
motor==3.6.0
pymongo==4.9.2
pydantic[email]==2.9.2
email-validator==2.2.0
python-multipart==0.0.9
bcrypt==4.0.1
